services/api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
cassandra-driver==3.29.2
python-dotenv==1.0.1
